pyproject.toml
[project]
name = "docforge-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datetime>=6.0",
    "document>=1.0",
    "docx>=0.2.4",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langchain-openai>=1.1.11",
    "langgraph>=1.0.10",
    "llm>=0.28",
    "notion>=0.1.0",
    "notion-client>=3.0.0",
    "numpy>=2.2.6",
    "openai>=2.26.0",
    "paragraph>=1.2.1",
    "platypus>=1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.9.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "regex>=2026.2.28",
    "reportlab>=4.4.10",
    "requests>=2.32.5",
    "responses>=0.26.0",
    "streamlit>=1.55.0",
    "uuid>=1.30",
    "uvicorn>=0.41.0",
]
